requirements.txt
langchain==0.1.12
langchain-experimental==0.0.54
chromadb==0.4.24
sentence_transformers==2.5.1
python-box
pysqlite3-binary==0.5.3
langchain-chroma
langchainhub 
Flask==3.0.3